pyproject.toml
[project]
name = "location-tool"
version = "0.1.0"
description = "餐厅智能搜索 CLI - 整合大众点评、小红书和高德地图数据"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
lt = "location_tool.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/location_tool"]
